[Feature #160589025] Apply review to update order status
