+++ server/controllers/orderControl.js
@@ -31,52 +31,19 @@ class orderControl {
     }
 
     static getOneOrder(req,resp) {
-        const order = orders.find(c => c.id == req.params.id);
-        if (!order) {
-            return resp.status(404).send({
-                status: 'error',
-                message: 'This order was not found on the list'
-            });
-        }else{
-            return resp.send({
-                status: 'success',
-                message: 'Returned one order',
-                order: order
-            });
-        }
+        return resp.send({
+            status: 'success',
+            message: 'Returned one order',
+            order: req.order
+        });
     }
 
     static updateOrderStatus(req,resp) {
-        const status = req.body.status;
-        const statusArray = ['pending','accepted','rejected','completed'];
-        const order = orders.find(c => c.id == req.params.id);
-        if (!order) {
-            return resp.status(404).send({
-                status: 'error',
-                message: 'This order was not found on the list'
-            });
-        }
-        if (status==="" || status==null){
-            return resp.status(400).send({
-                status: 'error',
-                message: 'Status field cannot be left empty'
-            });
-        }else if (validate.hasWhiteSpace(status) != false) {
-            return resp.status(400).send({
-                status: 'error',
-                message: 'No whitespaces allowed'
-            });
-        }else if (statusArray.includes(status) != true) {
-            return resp.status(400).send({
-                status: 'error',
-                message: 'The status you entered is invalid'
-            });
-        }
-        order.status = status;
+        req.order.status = req.status;
         return resp.send({
             status: 'success',
             message: 'Order status has been changed',
-            order: order
+            order: req.order
         });
     }
 }
