+++ server/routes/route.js
@@ -7,9 +7,9 @@ const Route = (app) => {
 
     app.post('/api/v1/orders', Middleware.validatePlaceOrder, orderControl.placeOrder);
 
-    app.get('/api/v1/orders/:id', orderControl.getOneOrder);
+    app.get('/api/v1/orders/:id', Middleware.validateGetOneOrder, orderControl.getOneOrder);
 
-    app.put('/api/v1/orders/:id', orderControl.updateOrderStatus);
+    app.put('/api/v1/orders/:id', Middleware.validateUpdateOrderStatus, orderControl.updateOrderStatus);
 }
 
 export default Route;
